Move proxy connection from OpenVPN service manager
